pathbuilder: simplify
