+++ go/vt/sqlparser/paths_test.go
@@ -17,10 +17,26 @@ limitations under the License.
 package sqlparser
 
 import (
+	"encoding/binary"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
 )
+
+// AddStep appends a single step (2 bytes) to path.
+func AddStep(path ASTPath, step ASTStep) ASTPath {
+	b := make([]byte, 2)
+	binary.BigEndian.PutUint16(b, uint16(step))
+	return path + ASTPath(b)
+}
+
+func AddStepWithOffset(path ASTPath, step ASTStep, offset int) ASTPath {
+	var buf [10]byte
+	binary.BigEndian.PutUint16(buf[0:], uint16(step))
+	n := binary.PutVarint(buf[2:], int64(offset))
+
+	return path + ASTPath(buf[:2+n])
+}
 
 func TestASTPathDebugString(t *testing.T) {
 	// select * from t where func(1,2,x) = 10
